requirements.txt
streamlit>=1.32.0
streamlit-extras>=0.4.0

# LLM & LangChain
langchain>=0.1.16
langchain-community>=0.0.38
langchain-core>=0.1.46
ollama>=0.1.8

# Retrieval
wikipedia-api>=0.6.0
duckdb>=0.10.1
chromadb>=0.4.24

# NLP / evaluation helpers
sentence-transformers>=2.7.0
transformers>=4.40.0
torch>=2.2.2

# Data handling
pandas>=2.2.1
numpy>=1.26.4
Pillow>=10.0.0
pydantic>=2.6.4

# Utilities
python-dotenv>=1.0.1
requests>=2.31.0
tenacity>=8.2.3
colorama>=0.4.6
PyYAML>=6.0.1
